(maint) Parse 'uuid.UUID' as string with format uuid
There's an issue when building the Java client library where it
interprets `UUID` as the Java type UUID, which then fails to import what
the API thinks is an array. [Based on this StackOverflow
thread](https://stackoverflow.com/questions/50204588/how-to-define-uuid-property-in-json-schema-and-open-api-oas)
it sounds like we want the UUID type to be interpreted as a String with
format `uuid`, which openapi will likely know how to build into the
correct type for libraries. This updates the parser to detect that
specific Go type and correctly type it in the API spec.